Add a trace sanity checker allocator
The intent is that we should be able to minimally check and "fix" any
trace we obtain so that later stages don't have to make decisions about
gaps in the trace and can just abort if things don't seem right.

Right now, this is a pretty shody, minimal initial effort.