sim/Clingy: rework revocation logic
Now supports revocation of not just the largest spans.  If given a
list of spans to revoke, it will pad the list out with the largest
spans, since it never makes sense to waste sweeping power.

TypelessClingy now supports a LRU queue of spans and can be told
to trade off its sweeping power between LRU and large spans.